refactored schedule components + fixed empty schedule error
